+++ pkg/therealssh/auth.go
