+++ pkg/therealssh/auth_test.go
