+++ pkg/therealssh/shell_darwin.go
